【【TBase】【2.15.19】dblink插件的DBLINK_COPY_TABLE不能返回拷贝的行数，增加返回的行数。】http://tapd.oa.com/pgxz/bugtrace/bugs/view?bug_id=1010092131088867115 (merge request !403)
Squash merge branch 'tbase_v2_hanway_616' into 'Tbase_v2.15.19'

* 【【TBase】【2.15.19】dblink插件的DBLINK_COPY_TABLE不能返回拷贝的行数，增加返回的行数。】http://tapd.oa.com/pgxz/bugtrace/bugs/view?bug_id=1010092131088867115

+++ contrib/dblink/dblink--1.2.sql
@@ -179,7 +179,7 @@ AS 'MODULE_PATHNAME', 'dblink_is_busy'
 LANGUAGE C STRICT PARALLEL RESTRICTED;
 
 CREATE FUNCTION dblink_copy_table(text, text, text, text, text)
-RETURNS int4
+RETURNS int8
 AS 'MODULE_PATHNAME', 'dblink_copy_table'
 LANGUAGE C STRICT PARALLEL RESTRICTED;
